requirements.txt
pygame==2.6.1
pyopengl==3.1.7
pyglm==2.7.3